Fix docker export
Nexus has a nexus.vmoptions file in the package bin directory which is owned by root and cannot be updated by the service
at runtime because it runs as the hab user. The only thing we actually need to modify in that file is the path to the Nexus
work directory which is normally /hab/svc/nexus/sonatype-work.

To overcome this we were creating a dangling symlink to a dynamically generated version of nexus.vmoptions in the
/hab/svc/nexus/config directory which only exists at run time. While this worked just fine for a hab environment running
directly in a VM, hab's container export function doesn't like it when we export to docker.

The fix is to remove dangling symlink and hard code the paths in the nexus.vmoptions file owned by root.

In reality, this should not be terrible, as it's still fairly easy to redirect the actual work directory using filesystem
mounts, or symlinks.

Signed-off-by: Richard Nixon <richard.nixon@btinternet>